카테고리 API 호출 api_control에 분리
